feat: implement numbers intelligence system with valuation models, portfolio scanning, and analytics dashboard

+++ backend/internal/model/clan.go
@@ -9,7 +9,7 @@ type Clan struct {
 	ChannelPhoto      string    `json:"channel_photo,omitempty"`
 	ChatTitle         string    `json:"chat_title"`
 	MembersCount      int       `json:"members_count"`
-	TotalScore        float64   `json:"total_score,omitempty"`
+	TotalScore        int64     `json:"total_score,omitempty"`
 	Rank              int       `json:"rank,omitempty"`
 	CreatedAt         time.Time `json:"created_at"`
 }
